Autocomplete: Discard laggy results.
Because the autocomplete API is asynchronous, sometimes we can receive
the nth result after the n+mth result. We then compute a
`lastReceivedTimestamp` variable representing the timestamp of the
received result. If a timestamp of a newly received result is lower than
this one, we can gently assume it is a laggy results and discard it.